fastapi-server-speakDocGPT/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
appnope==0.1.3
argilla==1.11.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
build==0.10.0
CacheControl==0.12.14
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
chromadb==0.3.26
cleo==2.0.1
click==8.1.3
clickhouse-connect==0.6.4
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
commonmark==0.9.1
compressed-rtf==1.0.6
crashtest==0.4.1
cryptography==41.0.1
dataclasses-json==0.5.9
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.6
dnspython==2.3.0
duckdb==0.8.1
dulwich==0.21.5
easygui==0.98.3
ebcdic==1.1.1
elevenlabs==0.2.18
email-validator==2.0.0.post2
et-xmlfile==1.1.0
executing==1.2.0
extract-msg==0.41.1
fastapi==0.99.1
filelock==3.12.2
Flask==2.3.2
Flask-Cors==3.0.10
flatbuffers==23.5.26
frozenlist==1.3.3
fsspec==2023.6.0
gpt4all==0.3.4
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
html5lib==1.1
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.15.1
humanfriendly==10.0
idna==3.4
IMAPClient==2.3.1
importlib-metadata==6.7.0
installer==0.7.0
ipython==8.14.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keyring==23.13.1
lanarky==0.7.12
langchain==0.0.221
langchainplus-sdk==0.0.19
lark-parser==0.12.0
llama-cpp-python==0.1.50
lockfile==0.12.2
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
msg-parser==1.2.0
msgpack==1.0.5
msoffcrypto-tool==5.0.1
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.0
olefile==0.46
oletools==0.60.1
onnxruntime==1.15.1
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
orjson==3.9.1
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandoc==2.3
parso==0.8.3
pcodedmp==1.2.6
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip==23.1.2
pkginfo==1.9.6
platformdirs==3.8.0
plumbum==1.8.2
ply==3.11
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
posthog==3.0.1
prompt-toolkit==3.0.38
protobuf==4.23.3
ptyprocess==0.7.0
pulsar-client==3.2.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.10
Pygments==2.15.1
PyMuPDF==1.22.3
pypandoc==1.11
pyparsing==2.4.7
pyproject_hooks==1.0.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
rapidfuzz==2.15.1
red-black-tree-mod==1.20
regex==2023.6.3
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.0.1
RTFDE==0.0.2
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==68.0.0
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
stack-data==0.6.2
starlette==0.27.0
stringcase==1.2.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.13.3
tomlkit==0.11.8
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
trove-classifiers==2023.5.24
typer==0.9.0
typing_extensions==4.7.1
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==4.2
ujson==5.8.0
unstructured==0.6.6
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
virtualenv==20.23.1
watchfiles==0.19.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.3.6
wheel==0.40.0
wrapt==1.14.1
xattr==0.10.1
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
